Fix (hopefully) syncing mail stuff again
The problem I was trying to fix before was that moving mail to the
inbox would immediately get moved out, since the post-new hook removes
the inbox tag from mail not in the inbox folder. Changing that though
results in mail archived elsewhere being put back in the inbox, since
it looks the same -- not in the inbox folder, but has the inbox tag.

Since in general we can't distinguish just by looking at the tag and
folder between "this mail wasn't in the Inbox, I did +inbox, so move
it there" and "this mail was moved out of the Inbox on the server,
so remove the inbox tag", we instead add a hook to the emacs notmuch
client to move mail into the inbox folder when adding the tag.