Focus listener to input unicode characters on macOS
Since the unicode input from the keyboard directly is awkward (it
requires a specific input method and doesn't work well with emoji).
Instead, the keyboardio now sends Focus events over the serial port
with the keycode; this little script, started as a launchagent,
listens, generates the corresponding character, and pastes it in via
system events.

A bit of a Rube Goldberg machine, but it seems to work...